ローダー関数の共通化
