+++ sample_frontend_container/sample_remix_react/app/routes/home.tsx
@@ -1,7 +1,8 @@
 import Header from '../components/Header';
 import Footer from '../components/Footer';
-import { LoaderFunction, ActionFunction, redirect } from '@remix-run/node';
+import { LoaderFunction, redirect } from '@remix-run/node';
 import { userDataLoader } from '../loader/userDataLoader';
+import { AuthenticationError } from '../utils/errors/AuthenticationError';
 
 /**
  * ローダー関数:
@@ -11,38 +12,19 @@ import { userDataLoader } from '../loader/userDataLoader';
  */
 export const loader: LoaderFunction = async ({ request }) => {
   try {
-    const userData = await userDataLoader(request);
-    // // HTTP-only クッキーの取得
-    // const cookieHeader = request.headers.get('Cookie');
-    // console.log('Loader: Incoming cookies:', cookieHeader);
+    const userData = await userDataLoader(request, false);
 
-    // // authToken と csrfToken をクッキーから抽出
-    // const authTokenMatch = cookieHeader?.match(/authToken=([^;]+)/);
-    // const csrfTokenMatch = cookieHeader?.match(/csrftoken=([^;]+)/);
-
-    // const authToken = authTokenMatch ? authTokenMatch[1] : null;
-    // const csrfToken = csrfTokenMatch ? csrfTokenMatch[1] : null;
-
-    // console.log('Loader: Extracted authToken:', authToken);
-    // console.log('Loader: Extracted csrfToken:', csrfToken);
-
-    // // authToken が存在しない場合はログインページへリダイレクト
-    // if (!authToken) {
-    //   console.log('Loader: Missing authToken, redirecting to login.');
-    //   return redirect('/login');
-    // }
-
-    // // 外部API呼び出し
-    // console.log('Loader: Fetching user data with authToken...');
-    // const userData = await fetchUserData(request);
-    // console.log('Loader: Retrieved user data:', userData);
-
-    // // 正常なレスポンスを返す
+    // 正常なレスポンスを返す
     return new Response(JSON.stringify(userData), {
       headers: { 'Content-Type': 'application/json' },
     });
   } catch (error) {
+    if (error instanceof AuthenticationError) {
+      console.log('Loader: AuthenticationError:');
+      return redirect('/login');
+    }
     console.error('Loader Error:', error);
+
     throw new Response('ユーザーデータの取得に失敗しました。', {
       status: 400,
     });
