+++ sample_frontend_container/sample_remix_react/app/utils/errors/AuthenticationError.tsx
@@ -1,0 +1,10 @@
+/**
+ * 認証エラーを表すカスタムエラークラス。
+ *
+ */
+export class AuthenticationError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'AuthenticationError';
+  }
+}
